scripts: Remove gcc-wrapper.py
Let's use the default output provided by the compiler and not use
the wrapper which adds extra dependencies (Python 3 in this case)
required to compile the kernel.

The wrapper also acted as a '-Werror' compiler flag and stopped the
compilation process on any warnings, in the upcoming changes we will
replace this by backporting the CONFIG_WERROR entry from the mainline
kernel.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>