dts: somc: tama: Do not re-configure touch reset GPIO
Do not reconfigure the touch reset GPIO, when the cont
splash screen is enabled, the SoMC Pinctrl driver will
reset the configuration set by the bootloader, and the
touchscreen will not be functional until the display driver
reconfigures the pin state. With the splash screen enabled,
this will only happen after the first power-off/on of the
display.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>