dts: somc: tama: Add dummy WLED device node to satisfy bootloader
The bootloader expects a specific backlight device node;
otherwise, it will turn off the backlight until the kernel
driver is initialized. In order for the cont splash screen
to function properly, with the bootloader not turning off
the backlight, we have no other choice but to specify this
device node.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>