dts: somc: tama: Fix the number of ADC-TM IIO channels
We already have a certain number of channels defined in the
MTP fragment. Here, we are adding additional ones, so the
list of IIO channels should be appended, not overwritten.
Otherwise, we will get an incorrect number of IIO channels
compared to the number of channel nodes.

Additionally, channel nodes from the MTP fragment were
added. This part of the change is not mandatory and was
made solely to avoid possible confusion.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>